removed user credential validation, fixed spelling of headers, created deletion validation sequence

+++ public/js/delete.js
@@ -1,9 +1,12 @@
 async function del(btn, snodeid){
-    fetch(`/api/codesnip/${snodeid}`, {
-        method: 'DELETE'
-    })
-    .then(res => {
-        console.log(res.text());
-        btn.closest(".snode").remove()
-    });
+
+    if (confirm('Are You Sure You Want to Delete This Snode?')) {
+        fetch(`/api/codesnip/${snodeid}`, {
+            method: 'DELETE'
+        })
+        .then(res => {
+            console.log(res.text());
+            btn.closest(".snode").remove()
+        });
+    }
 }
